refactor(CNV-43601): organize project into packages
Organize project code into packages. This improves
readability, maintainability and testability.
