+++ kubevirt-disk-uploader.yaml
@@ -52,11 +52,11 @@ spec:
             secretKeyRef:
               name: kubevirt-disk-uploader-credentials
               key: secretKey
-        - name: VM_NAMESPACE
+        - name: POD_NAMESPACE
           valueFrom:
             fieldRef:
               fieldPath: metadata.namespace
-        - name: VM_NAME
+        - name: POD_NAME
           valueFrom:
             fieldRef:
               fieldPath: metadata.name
